mcp: carve commands.rs (pure move)
handle_control_command_mcp — the control-socket/gateway ControlMsg
dispatch twin of the MCP tools — plus the control-result emitters and
start_task_with_state: 1,469 body lines and their 6 tests verbatim.
Named commands.rs because mod control would shadow crate::control (the
control socket) inside the mcp module.

Same glue as the controller_loop carve: use super::*, pub(crate)
widening, super::→crate:: path re-anchoring. Purity-checked; mcp::
tests 178/178.
